update category and record

+++ src/main/java/dev/usenkonastia/backend_lab2/service/CategoryService.java
@@ -4,10 +4,12 @@ import dev.usenkonastia.backend_lab2.domain.Category;
 import dev.usenkonastia.backend_lab2.repository.CategoryRepository;
 import dev.usenkonastia.backend_lab2.repository.UserRepository;
 import dev.usenkonastia.backend_lab2.service.exception.CategoryNotFoundException;
-import dev.usenkonastia.backend_lab2.service.exception.UserNotFoundException;
+import dev.usenkonastia.backend_lab2.service.exception.ForbiddenException;
 import dev.usenkonastia.backend_lab2.service.mapper.CategoryMapper;
 import jakarta.persistence.PersistenceException;
 import lombok.AllArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,6 +26,13 @@ public class CategoryService {
     @Transactional
     public Category addCategory(Category category) {
         try {
+            UUID currentUserId = getCurrentUser();
+            category = Category.builder()
+                    .categoryName(category.getCategoryName())
+                    .userId(currentUserId)
+                    .isPublic(category.getIsPublic())
+                    .records(category.getRecords())
+                    .build();
             return categoryMapper.toCategory(categoryRepository.save(categoryMapper.toCategoryEntity(category)));
         } catch (Exception e) {
             throw new PersistenceException(e);
@@ -46,10 +55,10 @@ public class CategoryService {
     }
 
     @Transactional(readOnly = true)
-    public List<Category> getUserCategories(UUID userId) {
+    public List<Category> getUserCategories() {
         try {
-            userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException(userId));
-            return categoryMapper.toCategoryList(categoryRepository.findByUserId(userId));
+            UUID currentUserId = getCurrentUser();
+            return categoryMapper.toCategoryList(categoryRepository.findByUserId(currentUserId));
         } catch (Exception e) {
             throw new PersistenceException(e);
         }
@@ -58,9 +67,33 @@ public class CategoryService {
     @Transactional
     public void deleteCategory(UUID id) {
         try {
+            boolean categoryExists = categoryRepository.existsById(id);
+            if (!categoryExists) {
+                return;
+            }
+            if (!doesHaveRights(id)) {
+                throw new ForbiddenException();
+            }
             categoryRepository.deleteById(id);
         } catch (Exception e) {
             throw new PersistenceException(e);
         }
+    }
+
+    private boolean doesHaveRights(UUID id) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        String email = auth.getName();
+        UUID currentUserId = userRepository.findByEmail(email).get().getId();
+        UUID userId = userRepository.findUserIdByCategoryId(id);
+        if (userId == null) {
+            throw new CategoryNotFoundException(id);
+        }
+        return userId.equals(currentUserId);
+    }
+
+    private UUID getCurrentUser() {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        String email = auth.getName();
+        return userRepository.findByEmail(email).get().getId();
     }
 }
