+++ src/main/java/dev/usenkonastia/backend_lab2/service/RecordService.java
@@ -5,16 +5,17 @@ import dev.usenkonastia.backend_lab2.entity.RecordEntity;
 import dev.usenkonastia.backend_lab2.repository.CategoryRepository;
 import dev.usenkonastia.backend_lab2.repository.RecordRepository;
 import dev.usenkonastia.backend_lab2.repository.UserRepository;
-import dev.usenkonastia.backend_lab2.service.exception.CategoryNotFoundException;
-import dev.usenkonastia.backend_lab2.service.exception.InvalidArgumentsException;
-import dev.usenkonastia.backend_lab2.service.exception.RecordNotFoundException;
-import dev.usenkonastia.backend_lab2.service.exception.UserNotFoundException;
+import dev.usenkonastia.backend_lab2.service.exception.*;
 import dev.usenkonastia.backend_lab2.service.mapper.RecordMapper;
 import jakarta.persistence.PersistenceException;
 import lombok.AllArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
 import java.util.List;
 import java.util.UUID;
 
@@ -39,11 +40,15 @@ public class RecordService {
     public Record addRecord(Record record) {
         try {
             RecordEntity recordEntity = recordMapper.toRecordEntity(record);
-            UUID userId = recordEntity.getUser().getId();
+            UUID userId = getCurrentUser();
             UUID categoryId = recordEntity.getCategory().getId();
-            userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException(userId));
+            record = Record.builder()
+                    .userId(userId)
+                    .categoryId(record.getCategoryId())
+                    .expense(record.getExpense())
+                    .date(ZonedDateTime.now(ZoneId.of("Europe/Kiev")))
+                    .build();
             categoryRepository.findById(categoryId).orElseThrow(() -> new CategoryNotFoundException(categoryId));
-
             return recordMapper.toRecord(recordRepository.save(recordMapper.toRecordEntity(record)));
         } catch (Exception e) {
             throw new PersistenceException(e);
@@ -53,6 +58,13 @@ public class RecordService {
     @Transactional
     public void deleteRecordById(UUID id) {
         try {
+            boolean isExist = recordRepository.existsById(id);
+            if (!isExist) {
+                return;
+            }
+            if (!doesHaveRights(id)) {
+                throw new ForbiddenException();
+            }
             recordRepository.deleteById(id);
         } catch (Exception e) {
             throw new PersistenceException(e);
@@ -69,5 +81,19 @@ public class RecordService {
             userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException(userId));
         }
         return recordMapper.toRecordList(recordRepository.findByUserIdAndCategoryId(userId, categoryId));
+    }
+
+    private UUID getCurrentUser() {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        String email = auth.getName();
+        return userRepository.findByEmail(email).get().getId();
+    }
+
+    private boolean doesHaveRights(UUID id) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        String email = auth.getName();
+        UUID currentUserId = userRepository.findByEmail(email).get().getId();
+        UUID userId = userRepository.findUserIdByRecordId(id);
+        return userId.equals(currentUserId);
     }
 }
